feat: create component that generates a static test type

+++ orogen_syskit_testsTypes.hpp
@@ -1,11 +1,14 @@
 #ifndef orogen_syskit_tests_TYPES_HPP
 #define orogen_syskit_tests_TYPES_HPP
 
+#include <stdint.h>
+
+
 namespace orogen_syskit_tests {
     struct Time {
-        int microseconds = 0;
-        int tv_sec = 0;
-        int tv_usec = 0;
+        int64_t microseconds = 0;
+        int64_t tv_sec = 0;
+        int64_t tv_usec = 0;
     };
 
     struct TestType {
